requirements.txt
requests
shodan
python-dateutil
python-whois
sherlock-project
crtsh